+++ core-ui/src/command-pallette/CommandPalletteUI/CommandPaletteUI.scss
@@ -1,4 +1,4 @@
-.compass-ui {
+.command-palette-ui {
   position: fixed;
   width: 100%;
   height: 100%;
