+++ core-ui/src/command-pallette/CommandPalletteUI/ResultsList/Result.js
@@ -29,7 +29,7 @@ export function Result({
   const actionText =
     typeof customActionText === 'string'
       ? customActionText
-      : t('compass.item-actions.navigate');
+      : t('command-palette.item-actions.navigate');
 
   return (
     <li
