+++ core-ui/src/command-pallette/CommandPalletteUI/ResultsList/ResultsList.scss
@@ -1,4 +1,4 @@
-.compass-ui__results {
+.command-palette-ui__results {
   max-height: 320px; // show ~5 items
   overflow: auto;
