+++ core-ui/src/command-pallette/CommandPalletteUI/components/components.scss
