+++ core-ui/src/command-pallette/CommandPalletteUI/handlers/helpers.js
