+++ core-ui/src/command-pallette/CommandPalletteUI/handlers/index.js
