+++ core-ui/src/command-pallette/CommandPalletteUI/handlers/nodesHandler.js
