+++ core-ui/src/command-pallette/CommandPalletteUI/handlers/tests/helpers.test.js
