+++ core-ui/src/command-pallette/CommandPalletteUI/useResourceCache.js
