+++ tests/cypress.json
@@ -55,7 +55,7 @@
     "certificates.spec.js",
     "test.settings.spec.js",
     "test-ingresses.spec.js",
-    "test-compass.spec.js",
+    "test-command-palette.spec.js",
     "run-after.spec.js"
   ]
 }
